requirements.txt
# Core dependencies (required for both session-based and persistent history)
mlflow>=2.21.2
streamlit==1.44.1
databricks-sdk

# Required for persistent history (uncomment if using persistent history)
# psycopg[binary,pool]>=3.1.0
